add demo for rpi281x lib
This file is from a prop I built a couple years ago when I was first learning to program. I wrote all the routines and copied everything from file to file, only changing the main() function to get the desired output, instead of removing superfluous routines. On the whole there's lots of bad practices, but the individual functions can serve as decent implementation examples.